requirements.txt
absl-py==0.12.0
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
btrfsutil==5.12.1
CacheControl==0.12.6
cachetools==4.2.2
caffeine-ng==3.5.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
chrome-gnome-shell==0.0.0
colorama==0.4.4
configobj==5.1.0.dev0
contextlib2==0.6.0.post1
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.9
decouple==0.0.7
defusedxml==0.7.1
dicom==0.9.9.post1
distlib==0.3.2
distro==1.5.0
dj-rest-auth==2.1.11
Django==3.2.6
django-bootstrap3==15.0.0
django-debug-toolbar==3.2.2
django-environ==0.4.5
django-filter==2.4.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.19
entrypoints==0.3
environ==1.0
ewmh==0.1.6
Faker==8.10.0
fasteners==0.15
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-api-core==1.28.0
google-api-python-client==2.5.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
gufw==21.4.0
h5py==3.1.0
html5lib==1.1
httplib2==0.19.1
idna==3.2
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
layoutswitcherlib==0.8.17
LibAppArmor==3.0.1
louis==3.18.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
migrate==0.3.8
mistune==0.8.4
monotonic==1.6
more-itertools==8.7.0
mpmath==1.2.1
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netsnmp-python==1.0a1
nose==1.3.7
notebook==6.3.0
npyscreen==4.10.5
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.2
pandas==1.2.3
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.5.29
ply==3.11
progress==1.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.16.0
psutil==5.8.0
ptyprocess==0.7.0
pulsectl==21.5.18
pwquality==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycairo==1.20.0
pycparser==2.20
pydicom==2.1.2
PyDrive==1.3.1
Pygments==2.9.0
PyGObject==3.40.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyrsistent==0.17.3
pysmbc==1.0.23
python-dateutil==2.8.1
python-decouple==3.4
python-libtorrent==1.2.11
python-xlib==0.30
pytz==2021.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
ranger-fm==1.9.3
reportlab==3.5.67
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
resolvelib==0.5.4
retrying==1.3.3
rsa==4.7.2
scipy==1.7.1
Send2Trash==1.5.0
setproctitle==1.2.2
shiboken2==5.15.2
six==1.16.0
soupsieve==2.2.1
speedtest-cli==2.1.3
sqlparse==0.4.1
sympy==1.8
team==1.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
tf-estimator-nightly==2.5.0.dev2021032501
tldextract==3.1.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ueberzug==18.1.9
ufw==0.36
uritemplate==3.0.1
urllib3==1.26.5
validate==5.1.0.dev0
virtualenv==20.4.7
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1
